Achieved working version of CassandraStore
Passes existing test suite (as copied in test_store)

pycassa is a blocking library (uses socket).
This is not compatible with the Twisted framework; the IStore interface
specifies the business methods be deferred operations, therefore, the
CassandraStore implementation returns defer.succeed to comply with the
IStore interface, even though all kvs calls in this context actually
completely block.
This is a fine compromise, for now.