Implement reactor.stop on AMQClient Close
This is not ideal (so far) but the Container process will now completly
exit whenever the AMQClient Closes.

The next step is to implement a container event (probably in the twisted
CapabilityContainer Service class, or something at that high level) that
can be called whenever a fatal container error occurs.

Note: The changes in this commit require using an updated txamqp carrot:
commit 95534860388a3219296a5ea7af6595b88cce5a56
A new version of carrot (0.10.12) will be released before this branch is
finalized.