Fixes a problem determining the git version.
When a non-ref'd commit was currently checked out, the code to determine the version would
die. This means that you cannot bisect between this current commit and when the change
was made to actively use this version determination (a85a813e055b9d364c2e279d1bfb68e5a12ff274).

To do a bisect here or after, you must cherry pick this commit on top or attempts to run things
will go awry.