Implement FatalError exception for crashing
The Container now exits under these circumstances:
 * AMQP Connect failure
 * AMQP Connector error (e.g. bad host name)
 * Ion App start error
 * When a FatalError exception is raised within a process op_ handler

Warning messages are logged from the container module, whenever a
fatalError event happens, and a detailed traceback if printed to stdout.
This will flash by if you are running a non-daemonized container and it
will be written to the twistd.log if you are running the container
daemonized.