PyCharm fixes. Some notes below:  Fixes OOIION-265
Need To Fix:
ion/util/ionlog.py:53: undefined name 'log_factory'
This is because of using the "standard" module level variables
definition pattern:
try:
    log_factory
except NameError:
    log_factory = LogFactory() #etc

NEW ATTEMPT:
if not hasattr(__name__, 'log_factory'):
    log_factory = LogFactory()

Nice To Fix:
ion/core/ioninit.py:39: local variable 'logconf' is assigned to but
never used
An if branch assigns this, and it is a reassignment of a variable that
is used. Incorrect output from pyflakes.

ion/core/cc/shell.py:119: 'identify' imported but unused
The purpose of these is to load them into the local namespace and then
extract them via locals()